update ped width condition in ped run
previously was failing at 24 adc counts. Turns out cell 0 has a fairly wide
pedetsal width in QHS, which was causing this to fail for almost every slot
(at least one channel, but often more). I've double the condition to 48
ADC counts, but this is worth looking into more carefully in the future.